logging and refactor
