+++ monitoring-service/src/main/resources/application.properties
@@ -22,9 +22,9 @@ preference.name.usage.node.gpu.delay=monitoring.gpu.usage.delay
 preference.name.usage.node.gpu.enable=monitoring.gpu.usage.enable
 preference.name.archive.runs.monitor.enable=monitoring.archive.runs.enable
 preference.name.archive.runs.monitor.delay=monitoring.archive.runs.delay
-preference.name.platform.usage.credit.monitor.enable=monitoring.platform.usage.credit.enable
-preference.name.platform.usage.credit.monitor.delay=monitoring.platform.usage.credit.delay
-platform.usage.credit.monitor.last.execution.file=lastPlatformUsageCreditMonitorTime.txt
+preference.name.platform.usage.credits.monitor.enable=monitoring.platform.usage.credits.enable
+preference.name.platform.usage.credits.monitor.delay=monitoring.platform.usage.credits.delay
+platform.usage.credits.monitor.last.execution.file=./lastPlatformUsageCreditsMonitorTime.txt
 preference.name.usage.node.pool.request.names=monitoring.node.pool.request.names
 
 # Elasticsearch
